patch for the encoder and interrupt
these 2 blocks will sent ascii 0 to the arduino, which is currently having issue in Scilab 6. This is a workaround to sent the +1 value to arduino, and v5 for arduino has been uploaded with a dummy input to counter this new index value.

Current workaround will try to minimize the changes on both Scilab and Arduino in order to be easily revert back to original when Scilab 6 fix the issue of ascii(0)